finished 4.21

+++ controllers/blogs.js
@@ -69,8 +69,25 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+
+        user.blogs = user.blogs.filter(b => b.toString() !== request.params.id.toString())
+        await user.save()
+        
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'token invalid' })
+    }
 })
 
 module.exports = blogsRouter
